introduce job

+++ src/FREE/Index/multigram_index.cpp
@@ -136,7 +136,7 @@ void free_index::MultigramIndex::insert_uni_bigram_into_index(
     //    else insert to expand
     std::unordered_set<char> uni_expand;
     for (const auto & [c, c_count] : unigrams) {
-        if (c_count/k_dataset_size_ <= k_threshold_) {
+        if (c_count/((double)k_dataset_size_) <= k_threshold_) {
             k_index_keys_.insert(std::string(1, c));
         } else {
             uni_expand.insert(c);
@@ -146,7 +146,7 @@ void free_index::MultigramIndex::insert_uni_bigram_into_index(
         // check if it is expand
         if (uni_expand.find(p.first) != uni_expand.end()) {
             std::string curr_str{p.first, p.second};
-            if (p_count/k_dataset_size_ <= k_threshold_) {
+            if (p_count/((double)k_dataset_size_) <= k_threshold_) {
                 k_index_keys_.insert(curr_str);
             } else {
                 expand.insert(curr_str);
@@ -161,7 +161,7 @@ void free_index::MultigramIndex::insert_kgram_into_index(
     // for each gram, if selectivity <= threshold, insert to index
     //    else insert to expand
     for (const auto & [s, s_count] : kgrams) {
-        if (s_count/k_dataset_size_ <= k_threshold_) {
+        if (s_count/((double)k_dataset_size_) <= k_threshold_) {
             k_index_keys_.insert(s);
         } else {
             expand.insert(s);
