+++ src/ngram_index.hpp
@@ -32,7 +32,7 @@ class NGramIndex {
     
     // the index structure should be stored here
     const std::vector<std::string> &k_dataset_;
-    const long double k_dataset_size_;
+    const unsigned int k_dataset_size_;
 
     std::set<std::string> k_index_keys_;
 };
